[BUGFIX] Specify mysql hostname in Settings.yaml
There was an issue, at least in the development vagrant box, where using
the default hostname '127.0.0.1' resulted in errors connecting to mysql.
This makes flow use 'localhost' instead of '127.0.0.1' to avoid that.

This also switches from php5-mysql to php5-mysqlnd. This is not related
to this bugfix, but it was too small to do a separate change.