[TASK] Use t3-mysql wrapper for mysql charset
This changes the mysql cookbook so that mysql uses utf8 by default.

We could either use t3-mysql or upgrade to a later version of the mysql
cookbook. Upgrading would be more involved, so just use t3-mysql for
now which wraps the same version of the mysql cookbook that we're
already using.